requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.7.4
pydantic-settings==2.4.0
httpx==0.27.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
streamlit==1.40.1
pandas==2.2.3
plotly==5.24.1
twilio>=9.0.0